Objects: Geometry版を追加。テストをひとつ追加。イベントの種類を拡充。

+++ Dev/Altseed.Reactive.Object/Common/INotifyUpdated.cs
@@ -9,7 +9,8 @@ namespace Altseed.Reactive.Object
 	{
 		/// <summary>
 		/// このオブジェクトシステムのインスタンスが更新されたときに値が流れるイベント。
+		/// 値として Engine.DeltaTime が流れます。
 		/// </summary>
-		IObservable<float> OnUpdateEvent { get; }
+		IObservable<long> OnUpdateEvent { get; }
 	}
 }
